requirements.txt
# Core deep learning framework
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Computer vision and model architectures
timm==0.9.12

# Data science and machine learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Image augmentation
albumentations==1.3.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
tqdm==4.66.1

# Configuration management
pyyaml==6.0.1

# Additional reproducibility tools
omegaconf==2.3.0
hydra-core==1.3.2
